+++ leptos-fluent/src/lib.rs
@@ -4,7 +4,8 @@
 //! [![License](https://img.shields.io/crates/l/leptos-fluent?logo=mit)](https://github.com/mondeja/leptos-fluent/blob/master/LICENSE.md)
 //! [![Tests](https://img.shields.io/github/actions/workflow/status/mondeja/leptos-fluent/ci.yml?label=tests&logo=github)](https://github.com/mondeja/leptos-fluent/actions)
 //! [![docs.rs](https://img.shields.io/docsrs/leptos-fluent?logo=docs.rs)][documentation]
-//! ![Crates.io downloads](https://img.shields.io/crates/d/leptos-fluent)
+//! [![Crates.io downloads](https://img.shields.io/crates/d/leptos-fluent)](https://crates.io/crates/leptos-fluent)
+//! [![Discord channel](https://img.shields.io/badge/-Discord-5865F2?style=flat-square&logo=discord&logoColor=white)](https://discord.com/channels/1031524867910148188/1251579884371705927)
 //!
 //! Internationalization framework for [Leptos] using [fluent-templates].
 //!
@@ -14,7 +15,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! leptos-fluent = "0.0.35"
+//! leptos-fluent = "0.0.37"
 //! fluent-templates = "0.9"
 //!
 //! [features]
@@ -25,6 +26,14 @@
 //!   "leptos-fluent/ssr",
 //!   "leptos-fluent/actix",  # actix and axum are supported
 //! ]
+//! ```
+//!
+//! If you're using `cargo-leptos` to build your project, watch the
+//! _locales/_ folder with:
+//!
+//! ```toml
+//! [package.metadata.leptos]
+//! watch-additional-files = ["locales"]  # Relative from Cargo.toml file
 //! ```
 //!
 //! # Usage
@@ -82,9 +91,12 @@
 //!
 //!         // Client side options
 //!         // -------------------
-//!         // Synchronize `<html lang="...">` attribute with the current
-//!         // language using `leptos::create_effect`. By default, it is `false`.
+//!         // Synchronize `<html lang="...">` attribute with the current active
+//!         // language. By default, it is `false`.
 //!         sync_html_tag_lang: true,
+//!         // Synchronize `<html dir="...">` attribute setting `"ltr"`,
+//!         // `"rtl"` or `"auto"` depending on the current active language.
+//!         sync_html_tag_dir: true,
 //!         // Update the language on URL parameter when using the method
 //!         // `I18n.set_language`. By default, it is `false`.
 //!         set_language_to_url_param: true,
@@ -103,6 +115,11 @@
 //!         // Get the initial language from `navigator.languages` if not
 //!         // found in the local storage. By default, it is `false`.
 //!         initial_language_from_navigator: true,
+//!         // Attributes to set for the language cookie. By default is `""`.
+//!         cookie_attrs: "SameSite=Strict; Secure; path=/; max-age=600",
+//!         // Update the language on cookie when using the method `I18n.set_language`.
+//!         // By default, it is `false`.
+//!         set_language_to_cookie: true,
 //!
 //!         // Server side options
 //!         // -------------------
@@ -117,9 +134,6 @@
 //!         cookie_name: "lang",
 //!         // Get the initial language from cookie. By default, it is `false`.
 //!         initial_language_from_cookie: true,
-//!         // Update the language on cookie when using the method `I18n.set_language`.
-//!         // By default, it is `false`.
-//!         set_language_to_cookie: true,
 //!         // URL parameter name to use discovering the initial language
 //!         // of the user. By default is `"lang"`.
 //!         url_param: "lang",
@@ -201,12 +215,6 @@
 //! - [Examples]
 //! - [Documentation]
 //!
-//! # Roadmap
-//!
-//! Leptos-fluent is currently ready for most use cases. However, it is still in an
-//! early stage of development and the API may contain breaking changes through
-//! v0.0.X releases.
-//!
 //! [leptos]: https://leptos.dev/
 //! [fluent-templates]: https://github.com/XAMPPRocky/fluent-templates
 //! [quickstart]: https://docs.rs/leptos-fluent/latest/leptos_fluent/macro.leptos_fluent.html
@@ -234,8 +242,33 @@ use leptos::{
 };
 pub use leptos_fluent_macros::leptos_fluent;
 
+#[doc(hidden)]
+#[cfg_attr(debug_assertions, derive(Debug))]
+pub enum WritingDirection {
+    Ltr,
+    Rtl,
+    Auto,
+}
+
+impl WritingDirection {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            WritingDirection::Ltr => "ltr",
+            WritingDirection::Rtl => "rtl",
+            WritingDirection::Auto => "auto",
+        }
+    }
+}
+
+impl core::fmt::Display for WritingDirection {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        f.write_str(self.as_str())
+    }
+}
+
 /// Each language supported by your application.
-#[derive(Clone, Debug)]
+#[derive(Clone)]
+#[cfg_attr(debug_assertions, derive(Debug))]
 pub struct Language {
     /// Language identifier
     ///
@@ -246,6 +279,8 @@ pub struct Language {
     /// The name of the language, such as `English`, `Español`, etc.
     /// This name will be intended to be displayed in the language selector.
     pub name: &'static str,
+    /// Writing direction of the language
+    pub dir: &'static WritingDirection,
 }
 
 impl Language {
@@ -323,6 +358,16 @@ pub struct I18n {
     pub languages: &'static [&'static Language],
     /// Static translations loader of fluent-templates.
     pub translations: ReadSignal<Vec<&'static Lazy<StaticLoader>>>,
+}
+
+#[cfg(debug_assertions)]
+impl core::fmt::Debug for I18n {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        f.debug_struct("I18n")
+            .field("language", &self.language.get())
+            .field("languages", &self.languages)
+            .finish()
+    }
 }
 
 /// Get the current context for localization.
@@ -440,8 +485,13 @@ pub fn language_from_str_between_languages(
         {
             Some(lang) => Some(lang),
             None => {
-                let mut lazy_target_lang = target_lang.clone();
-                lazy_target_lang.region = None;
+                let lazy_target_lang =
+                    LanguageIdentifier::from_raw_parts_unchecked(
+                        target_lang.language,
+                        None,
+                        None,
+                        None,
+                    );
                 match languages
                     .iter()
                     .find(|lang| lang.id.matches(&lazy_target_lang, true, true))